VMP_Parser/backend/requirements.txt
streamlit
pytesseract
opencv-python
pillow
googletrans==4.0.0-rc1
gtts
python-dotenv
google-generativeai
